+++ PeteBJJCompFeed/PeteBJJCompFeediOS/Feed UI/Composers/CompetitionsUIComposer.swift
@@ -29,38 +29,6 @@ public final class CompetitionsUIComposer {
     }
 }
 
-private final class MainQueueDispatchDecorator<T> {
-    private let decoratee: T
-    
-    init(decoratee: T) {
-        self.decoratee = decoratee
-    }
-    
-    func dispatch(completion: @escaping () -> Void) {
-        guard Thread.isMainThread else {
-            return DispatchQueue.main.async(execute: completion)
-        }
-        
-        completion()
-    }
-}
-
-extension MainQueueDispatchDecorator: CompetitionsLoader where T == CompetitionsLoader {
-    func load(completion: @escaping (CompetitionsLoader.Result) -> Void) {
-        decoratee.load { [weak self] result in
-            self?.dispatch { completion(result) }
-        }
-    }
-}
-
-extension MainQueueDispatchDecorator: EventImageDataLoader where T == EventImageDataLoader {
-    func loadImageData(from url: URL, completition: @escaping (EventImageDataLoader.Result) -> Void) -> any EventImageDataLoaderTask {
-        return decoratee.loadImageData(from: url) { [weak self] result in
-            self?.dispatch { completition(result) }
-        }
-    }
-}
-
 private extension CompetitionsViewController {
     static func makeWith(delegate: CompetitionsViewControllerDelegate, title: String) -> CompetitionsViewController {
         let bundle = Bundle(for: CompetitionsViewController.self)
